refactor(web): Isolate module "CommandCenter"
