+++ ui/web/src/modules/CommandCenter/index.ts
@@ -1,0 +1,1 @@
+export { CommandCenter, executeCommand, registerCommand } from './CommandCenter';
