+++ ui/web/src/modules/Terminals/TerminalList.tsx
@@ -4,7 +4,7 @@ import { useObservableState } from 'observable-hooks';
 import React from 'react';
 import { defer, repeat, retry, shareReplay, switchMap, toArray } from 'rxjs';
 import { terminal$ } from '../../common/create-connection';
-import { registerCommand } from '../CommandCenter/CommandCenter';
+import { registerCommand } from '../CommandCenter';
 import { openPage } from '../Pages';
 import { TerminalListItem } from './TerminalListItem';
