+++ ui/web/src/modules/TradeCopier/TradeCopyRelationList.tsx
@@ -6,7 +6,7 @@ import { useObservable, useObservableState } from 'observable-hooks';
 import React, { useState } from 'react';
 import { combineLatest, first, mergeMap, tap, toArray } from 'rxjs';
 import { terminal$ } from '../../common/create-connection';
-import { registerCommand } from '../CommandCenter/CommandCenter';
+import { registerCommand } from '../CommandCenter';
 import Form from '../Form';
 import { openPage } from '../Pages';
 import { terminate } from '../Terminals/TerminalListItem';
