+++ ui/web/src/modules/User/UserMenu.tsx
@@ -15,7 +15,7 @@ import { useObservableState } from 'observable-hooks';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { authState$, supabase } from '../../common/supabase';
-import { executeCommand, registerCommand } from '../CommandCenter/CommandCenter';
+import { executeCommand, registerCommand } from '../CommandCenter';
 import i18n from '../Locale/i18n';
 import { currentHostConfig$ } from '../Workbench/model';
